requirements.txt
tweepy
python-dotenv
kaleido
datetime
pytz